requirements.txt
flask
flask-mysql-connector
flask-restful
mysql-connector-python
pandas
python-dateutil
tzdata
